requirements.txt
requests
bs4
lxml
textblob
transformers
torch
openai
python-dotenv